Move all constants into a single module
